+++ src/Nav_Content/index.d.ts
@@ -1,0 +1,5 @@
+import type { Ctx } from '@ctx-core/object'
+import type { SvelteComponentTyped } from 'svelte'
+export class Nav_Content extends SvelteComponentTyped<{
+	ctx:Ctx
+}> {}