+++ src/Nav_Content/index.js
@@ -1,0 +1,2 @@
+import 'svelte'
+export * as Nav_Content from './index.svelte'