+++ src/Nav_Header/index.d.ts
@@ -1,0 +1,5 @@
+import type { Ctx } from '@ctx-core/object'
+import type { SvelteComponentTyped } from 'svelte'
+export class Nav_Header extends SvelteComponentTyped<{
+	ctx:Ctx
+}> {}