+++ src/Nav_Header/index.js
@@ -1,0 +1,2 @@
+import 'svelte'
+export * as Nav_Header from './index.svelte'