+++ src/Nav_Header/index.svelte
@@ -1,0 +1,20 @@
+<script>
+import { nav_opened__close } from '@ctx-core/nav'
+/** @type {import('@ctx-core/object').Ctx} */
+export let ctx
+</script>
+
+<div class="Nav_Header header">
+	<slot name="close">
+		<a href="." class="close" on:click|preventDefault={()=>nav_opened__close(ctx)}>
+			<slot name="in__close">×</slot>
+		</a>
+	</slot>
+	<slot></slot>
+</div>
+
+<style>
+.Nav_Header a {
+	text-decoration: none;
+}
+</style>