+++ src/Nav_Item/index.d.ts
@@ -1,0 +1,9 @@
+import { SvelteComponentTyped } from 'svelte'
+export class index extends SvelteComponentTyped<{
+	selected?:boolean
+	class?:string
+	href?:string
+	alt?:string
+	title?:string
+	a_class?:string
+}> {}