+++ src/index.d.ts
@@ -1,0 +1,5 @@
+import 'svelte'
+export * as Nav_Content from './Nav_Content'
+export * as Nav_Handle from './Nav_Handle'
+export * as Nav_Header from './Nav_Header'
+export * as Nav_Item from './Nav_Item'